# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: AuctioNet API
  version: 0.1.0
servers:
  - url: http://localhost:5000
  - url: http://localhost:8080
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id: { type: string, example: u_1 }
        email: { type: string, format: email }
        balance: { type: number, example: 100000 }
        held: { type: number, example: 0 }
        purchases:
          type: array
          items: { type: string, example: p_1 }
    Product:
      type: object
      properties:
        id: { type: string, example: p_1 }
        owner_id: { type: string, example: u_1 }
        title: { type: string }
        description: { type: string }
        category: { type: string, example: vehicule }
        images:
          type: array
          items: { type: string, example: media/p_1/20250101T120000Z_img.jpg }
    Auction:
      type: object
      properties:
        id: { type: string, example: a_1 }
        product_id: { type: string, example: p_1 }
        start_price: { type: number, example: 5000 }
        min_increment: { type: number, example: 50 }
        start_at: { type: string, format: date-time }
        end_at: { type: string, format: date-time }
        status: { type: string, enum: [scheduled, running, closed] }
        current_price: { type: number, example: 5100 }
        current_bid_id: { type: [string, 'null'], example: b_3 }
        product:
          $ref: '#/components/schemas/Product'
    BidResult:
      type: object
      properties:
        ok: { type: boolean, example: true }
        bid_id: { type: string, example: b_3 }
        current_price: { type: number, example: 5150 }

paths:
  /api/health:
    get:
      summary: Health check
      responses:
        '200': { description: OK }

  /api/categories:
    get:
      summary: Liste des catégories
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  categories:
                    type: array
                    items: { type: string, example: vehicule }

  /api/auth/register:
    post:
      summary: Inscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 6 }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string }
                  email: { type: string }
                  balance: { type: number }

  /api/auth/login:
    post:
      summary: Connexion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        '200':
          description: Token JWT
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token: { type: string }
        '401': { description: Invalid credentials }

  /api/me:
    get:
      summary: Profil courant
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }

  /api/products:
    get:
        summary: Liste des produits
        parameters:
        - in: query
          name: owner_id
          schema: { type: string }
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: search
          schema: { type: string }
        responses:
          '200':
            description: OK
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    categories:
                      type: array
                      items: { $ref: '#/components/schemas/Product' }
    post:
      summary: Créer un produit
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, description, category]
              properties:
                title: { type: string }
                description: { type: string }
                category: { type: string, example: vehicule }
                images:
                  type: array
                  items: { type: string }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string }

  /api/products/{pid}/images:
    post:
      summary: Upload d'image produit
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: pid
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Stored
          content:
            application/json:
              schema:
                type: object
                properties:
                  path: { type: string, example: media/p_1/20251105T120000Z_img.jpg }
        '400': { description: Bad Request (type/poids) }
        '403': { description: Not owner }
        '404': { description: Product not found }

  /api/auctions:
    get:
      summary: Lister les enchères
      parameters:
        - in: query
          name: status
          schema: { type: string, enum: [scheduled, running, closed] }
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: search
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  auctions:
                    type: array
                    items: { $ref: '#/components/schemas/Auction' }
    post:
      summary: Créer une enchère
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [product_id, start_price, start_at, end_at]
              properties:
                product_id: { type: string }
                start_price: { type: number }
                start_at: { type: string, format: date-time }
                end_at: { type: string, format: date-time }
                min_increment: { type: number, example: 50 }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Auction' }

  /api/auctions/{id}:
    get:
      summary: Détail d'une enchère
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Auction' }
        '404': { description: Not found }

  /api/auctions/{id}/bids:
    post:
      summary: Placer une mise
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount]
              properties:
                amount: { type: number, example: 5200 }
      responses:
        '201':
          description: Placed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BidResult' }
        '400': { description: Business rule failed }

  /api/my/purchases:
    get:
      summary: Produits achetés
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  products:
                    type: array
                    items: { $ref: '#/components/schemas/Product' }

  /media/{pid}/{filename}:
    get:
      summary: Servir une image (publique)
      parameters:
        - in: path
          name: pid
          required: true
          schema: { type: string }
        - in: path
          name: filename
          required: true
          schema: { type: string }
      responses:
        '200': { description: Image bytes }
        '404': { description: Not found }